+++ CHANGELOG.md
@@ -4,10 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.0.3]
+### Changed
+- Moved output files to optional arguments.
+### Added
+- `-s` option to add in a `sample_name` column to outputs.
 
 ## [v0.0.2]
 ### Changed
-- No end-user changes
+- No end-user changes.
 
 ## [v0.0.1]
 ### Added